Learning glottis segmentation from HLE Dataset

+++ dataset.py
@@ -240,4 +240,4 @@ class HLEOnlyGlottis(Dataset):
             image = augmentations["image"]
             segmentation = augmentations["masks"][0]
 
-        return image, segmentation
+        return image, segmentation.int()
